requirements.txt
kivy>=2.3.0
buildozer>=1.5.0
kivmob>=2.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
